pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smartapi-test"
version = "0.1.0"
description = "智能声明式 API 自动化测试平台"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SmartAPI Team"},
]
dependencies = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "pytest-html>=4.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "jsonpath-ng>=1.6",
    "lxml>=5.0",
    "jinja2>=3.1",
    "click>=8.1",
    "rich>=13.0",
    "faker>=22.0",
    "allure-pytest>=2.13",
    "cryptography>=42.0",
    "mcp>=1.0",
    "pydantic>=2.5",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2",
    "mypy>=1.8",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
]
web = [
    "fastapi>=0.109",
    "uvicorn>=0.27",
]
all = [
    "smartapi-test[dev,web]",
]

[project.scripts]
smartapi = "smartapi.cli.main:cli"

[project.entry-points.pytest11]
smartapi = "smartapi.pytest_plugin"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "smoke: Smoke tests",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.hatch.build.targets.wheel]
packages = ["smartapi"]
